Refactor user, order, and enquiry blade templates
Simplify the code in the user.blade.php, order.blade.php, and enquiry.blade.php templates by removing unnecessary PHP code and replacing it with the Laravel Blade syntax. This improves readability and maintainability of the templates.

Deleted CheckWebUser middleware

Removed the CheckWebUser middleware as it is no longer needed in the application.

Update login.blade.php

Commented out the "Forgot Password?" link in the login.blade.php template.

Update transaction.blade.php

Simplify the code in the transaction.blade.php template by removing unnecessary PHP code and replacing it with the Laravel Blade syntax. This improves readability and maintainability of the template.

Update create_gsts_table migration

Added two new columns, 'status' and 'deleteId', to the create_gsts_table migration. The 'status' column is a boolean column with a default value of 1, and the 'deleteId' column is a boolean column with a default value of 0. This allows for better management of GST records in the application.

Update models and migrations

Performed updates to the models and migrations in the repository.